Move country sorting from JavaScript to Python
The sorting of countries for selects in forms like join, new member and edit
member was done in JavaScript depending on an by now old jQuery version causing
problems with the new jQuery version.

In order to solve this issue and remove the dependency completely, the sorting
is moved to Python. It is still performed depending on the language chosen by
the user.