Send emails from yes@c3s.cc instead of noreply
It is impolite to send an email from noreply@c3s.cc. In case there is any issue
with the email sent like a link is not working or the recipient having a
question about the matter, they needt to get in contact anyway. Intentionally
denying them the possibility to reply to an email make life harder for
everyone. Instead, an email address should be used which helps reaching someone
who can take care of the matter.