Simplify page size change
The form to change the page size, i.e. elements displayed per page, took to
much space for that it was not very important. Instead, the number of elements
per page is not displayed in the same row as the pagination arrows aligned to
the right. The number of elements is a link and when clicked a dialog is opened
which contains the form to change the number of elements per page. The input is
focussed and selected.