Show message for empty general assembly list in member detail page
General assemblies are only relevant for a member if their date is after the
membership was approved. Thus, in case there is no general assembly planned for
a member when they are approved, the list of general assemblies relevant for
this member is empty. Instead of showing an empty table, a message is shown
that there are not relevant general assemblies for this member.