Separate dues calculation logic
Dues calculation logic is moved to a separate class in the business logic
layer and applied to dues 2019.

There are two things still not ideal. First, the methods calculate_quarter and
get_description of the QuarterlyDuesCalculator class should be private but are
currently needed in the dues_2019 module. Second, the description should not be
part of the message text but instead be a line item description, stored with
the dues and taken from the dues when used later on.