Fix membership listing validation
The maximum and default date of the listing by date form and as well the year
list of the listing by year form were generated on application start and not
dynamically during form rendering. This led to the fact that with the turn of
the year from 2019 to 2020 the new year did not appear in the year list and the
current date could not be entered because it was considered in the past. Now
the dates are handled dynamically during form rendering.