Further improve currency amount formatting
Apparently, there are problems in some circumstances with locale settings of
the application causing the non-breaking space and Euro sign to be placed in
the beginning or the end of the string. Thus, now the locale is set for
formatting and the characters are removed instead of the string truncated.